[+] Added [-] Removed [*] Edited
[+] Added TestCommand
[+] Added JoinListener

[*] Set all long Objects to int

+++ src/main/java/net/marakaner/economysystem/money/BankAccount.java
@@ -2,13 +2,13 @@ package net.marakaner.economysystem.money;
 
 public class BankAccount {
 
-    private long money;
+    private int money;
 
-    public BankAccount(long money) {
+    public BankAccount(int money) {
         this.money = money;
     }
 
-    public long getMoney() {
+    public int getMoney() {
         return money;
     }
 }
